requirements.txt
selenium
pytest
pytest-selenium
allure-pytest
python-dotenv